Start testing against Node 18 too
Requires a workaround due to IPFS using Node's new fetch() global in
Node 18, but this should be fixed soon both in Undici and in Mockttp
(which should be setting an explicit content-length in simple cases like
this).

Using transfer-encoding: chunked alone isn't enough, unfortunately. In
that case, with Mockttp 3.0.0, we still drop default connection
behaviour, and this results in Mockttp closing immediately after the
response, but Undici assuming the connection is still open, creating a
race condition that reliably breaks things and kills the 2nd Undici
request. For now we set keep-alive (triggering node's KA) and this will
hopefully be fixed by Undici soon, and tangentially also fixed by the
Mockttp v3.0.1 change to come too.