Allow proxying unmatched requests to a real IPFS node
This requires dropping Node v14 testing, and so implicitly dropping
official support, because IPFS itself doesn't run when running Node <16
which makes testing really really awkward, even though MockIPFS runs
fine in that environment. Since it's unsupported by IPFS though, seems
reasonable to match their requirements. This is a breaking change but
we're still pre-1.0 so that's not such a big deal.